kern compiler fixes swapbuf tweaks mklocale updated

+++ usr.bin/mklocale/ldef.h
@@ -40,14 +40,14 @@
  * This should look a LOT like a _RuneEntry
  */
 typedef struct rune_list {
-    rune_t				min;
-    rune_t 				max;
-    rune_t 				map;
-    u_long				*types;
-    struct rune_list	*next;
+    rune_t			min;
+    rune_t 			max;
+    rune_t 			map;
+    u_int32_t			*types;
+    struct rune_list		*next;
 } rune_list;
 
 typedef struct rune_map {
-    u_long				map[_CACHED_RUNES];
+    u_int32_t			map[_CACHED_RUNES];
     rune_list			*root;
 } rune_map;
